wip: move main header code to folder; add potential 'minimal' style
