+++ src/app/shared/shared.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from "@angular/core";
 import { CommonModule } from "@angular/common";
 import { IonicModule } from "@ionic/angular";
-import { PLHMainHeaderComponent } from "./components/plh-main-header";
+import { PLHMainHeaderComponent } from "./components/header/header.component";
 import { RouterModule } from "@angular/router";
 import { TemplateComponentsModule } from "./components/template/template.module";
 import { ReactiveFormsModule } from "@angular/forms";
